rpi3: Load the vc4 module before starting psplash
If we do not load the vc4 module at the early stage, then the
module will be loaded by udev. When loading the vc4 module,
the graphical mode is changed. If psplash is already running,
the displaying of the psplash boot process will be disabled
and a blank screen will be displayed.

Signed-off-by: Anton Kikin <a.kikin@tano-systems.com>